requirements.txt
asgiref==3.5.2
Django==4.0.6
django-appconf==1.0.5
django-filter==22.1
django-select2==7.10.0
geographiclib==1.52
geopy==2.2.0
Pillow==9.2.0
sqlparse==0.4.2
